+++ CHANGELOG.md
@@ -110,6 +110,7 @@ check of rewards
 command to modify `max-finality-providers` parameter.
 - [#1344](https://github.com/babylonlabs-io/babylon/pull/1344) Add `MsgType` field to `QueuedMessageResponse` message.
 - [#1352](https://github.com/babylonlabs-io/babylon/pull/1352) Add stake expansion signatures in BTC delegation query response
+- [#1358](https://github.com/babylonlabs-io/babylon/pull/1358) Simplify selective slashing message handler
 - [#1367](https://github.com/babylonlabs-io/babylon/pull/1367) Add v3 mainnet upgrade handler
 - [#1381](https://github.com/babylonlabs-io/babylon/pull/1381) Add multi-staking allow-list logic